requirements.txt
flask==3.0.3
flask-cors==4.0.0
Flask-SocketIO==5.3.6
eventlet==0.36.1
python-dotenv==1.0.1
pymongo[srv]==4.8.0
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
openai==1.51.0
requests==2.32.3
gunicorn==22.0.0



